+++ handlers/student-details.go
@@ -11,6 +11,27 @@ import (
 )
 
 func (h *Handler) InsertCAd(ctx *gin.Context) {
+
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
+
 	var cad models.StudentInfo
 	err := ctx.BindJSON(&cad)
 	if err != nil {
@@ -26,7 +47,6 @@ func (h *Handler) InsertCAd(ctx *gin.Context) {
 
 func (h *Handler) RetrieveValuesCAd(ctx *gin.Context) {
 
-	//token := ctx.Param("token")
 	token, err1 := ctx.Cookie("token")
 	if err1 != nil {
 		ctx.JSON(http.StatusInternalServerError, err1.Error())
@@ -56,6 +76,26 @@ func (h *Handler) RetrieveValuesCAd(ctx *gin.Context) {
 
 func (h *Handler) UpdateValuesCAd(ctx *gin.Context) {
 
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
+
 	oldCourse := ctx.Param("coursename")
 	var rcd models.StudentInfo
 	ctx.BindJSON(&rcd)
@@ -70,6 +110,26 @@ func (h *Handler) UpdateValuesCAd(ctx *gin.Context) {
 
 func (h *Handler) DeleteSA(ctx *gin.Context) {
 
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
+
 	rollNumber := ctx.Param("rollnumber")
 
 	err := h.service.DeleteStudent(rollNumber)
@@ -81,6 +141,27 @@ func (h *Handler) DeleteSA(ctx *gin.Context) {
 
 }
 func (h *Handler) UpdateStudentNameAndAge(ctx *gin.Context) {
+
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
+
 	var rcd *models.StudentInfo
 	existing_name := ctx.Param("name")
 	ctx.BindJSON(&rcd)
@@ -125,6 +206,26 @@ func (h *Handler) FetchAllCourseForAStudent(ctx *gin.Context) {
 }
 
 func (h *Handler) DeleteStudentCourse(ctx *gin.Context) {
+
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
 
 	parameter := ctx.Params
 	student_name := parameter.ByName("name")
